refactor: update project structure (#6)
