+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+export * from "./Catalog";
+export * from "./MealPlans";
+export * from "./Recipes";
+export * from "./ShoppingList";
