+++ src/queries/shopping-list.graphql.ts
