--- src/views/catalog/Catalog/index.ts
@@ -1,2 +1,0 @@
-export * from "./Catalog";
-export * from "./query.graphql";
