--- src/views/catalog/components/AddCatalogItemForm/index.ts
@@ -1,1 +1,0 @@
-export * from "./AddCatalogItemForm";
