--- src/views/catalog/index.ts
@@ -1,2 +1,0 @@
-export * from "./Catalog";
-export * from "./components";
