--- src/views/index.ts
@@ -1,5 +1,0 @@
-export * from "./item-inventory";
-export * from "./meal-plans";
-export * from "./recipes";
-export * from "./shopping-list";
-export * from "./catalog";
