--- src/views/item-inventory/ItemInventory/ItemInventory.tsx
@@ -1,8 +1,0 @@
-import { Typography } from "@mui/material";
-import { Layout } from "components";
-
-export const ItemInventory = () => (
-  <Layout>
-    <Typography>Item Inventory</Typography>
-  </Layout>
-);
