--- src/views/item-inventory/ItemInventory/index.ts
@@ -1,1 +1,0 @@
-export * from "./ItemInventory";
