--- src/views/item-inventory/index.ts
@@ -1,1 +1,0 @@
-export * from "./ItemInventory";
