--- src/views/meal-plans/MealPlans/index.ts
@@ -1,1 +1,0 @@
-export * from "./MealPlans";
