--- src/views/recipes/Recipes/index.ts
@@ -1,1 +1,0 @@
-export * from "./Recipes";
