--- src/views/shopping-list/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./List";
-export * from "./AddShoppingItemForm";
