--- src/views/shopping-list/index.ts
@@ -1,3 +1,0 @@
-export * from "./ShoppingList";
-export * from "./components";
-export * from "./query.graphql";
